+++ src/googleTTS/GoogleTTSRequest.ts
@@ -2,13 +2,13 @@ import { CommonTTSRequest } from '../common/commonTTSRequest';
 
 export class GoogleTTSRequest extends CommonTTSRequest {
     text: string;
-    lang = 'en';
+    lang = 'en-US';
     slow = false;
 
     constructor(text: string, lang?: string) {
         super();
         this.text = text;
-        this.lang = lang || 'en';
+        this.lang = lang || 'en-US';
     }
 
     summery(): string {